FastAPI-Mongo/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
bidict==0.22.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-widget-tweaks==1.5.0
dnspython==2.6.0
exceptiongroup==1.2.0
fastapi==0.109.2
Flask==2.3.3
Flask-SocketIO==5.3.6
h11==0.14.0
httptools==0.6.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.7.1
python-socketio==5.9.0
PyYAML==6.0.1
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing_extensions==4.8.0
tzdata==2023.3
unicorn==2.0.1.post1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.7
wsproto==1.2.0
